+++ packages/react/src/components/Radio/RadioGroup/index.css
@@ -1,0 +1,5 @@
+.charcoal-radio-group {
+  display: grid;
+  grid-template-columns: 1fr;
+  grid-gap: 8px;
+}
