+++ packages/react/src/components/Radio/index.css
@@ -26,9 +26,11 @@
   background-color: var(--charcoal-surface1);
   transition: 0.2s background-color, 0.2s box-shadow;
 }
+
 .charcoal-radio__input:checked {
   background-color: var(--charcoal-brand);
 }
+
 .charcoal-radio__input:checked::after {
   content: '';
   display: block;
@@ -56,13 +58,16 @@
 .charcoal-radio__input:not(:disabled):active {
   background-color: var(--charcoal-surface1-press);
 }
+
 .charcoal-radio__input:not(:disabled):focus {
   outline: none;
   box-shadow: 0 0 0 4px rgba(0, 150, 250, 0.32);
 }
+
 .charcoal-radio__input:not(:disabled):focus-visible {
   box-shadow: 0 0 0 4px rgba(0, 150, 250, 0.32);
 }
+
 .charcoal-radio__input:not(:disabled):focus:not(:focus-visible) {
   box-shadow: none;
 }
@@ -75,12 +80,15 @@
 .charcoal-radio__input:checked:not(:disabled):hover {
   background-color: var(--charcoal-brand-hover);
 }
+
 .charcoal-radio__input:checked:not(:disabled):hover::after {
   background-color: var(--charcoal-text5-hover);
 }
+
 .charcoal-radio__input:checked:not(:disabled):active {
   background-color: var(--charcoal-brand-press);
 }
+
 .charcoal-radio__input:checked:not(:disabled):active::after {
   background-color: var(--charcoal-text5-press);
 }
@@ -89,14 +97,4 @@
   font-size: 14px;
   line-height: 22px;
   color: var(--charcoal-text2);
-}
-
-.charcoal-radio-group {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-gap: 8px;
-}
-
-.charcoal-radio-group[aria-disabled='true'] {
-  opacity: 0.32;
 }
