+++ packages/react/src/components/Radio/index.tsx
@@ -1,9 +1,11 @@
-import { memo, forwardRef, useCallback, useContext, useMemo } from 'react'
+import './index.css'
+
+import { memo, forwardRef, useCallback, useContext } from 'react'
 import * as React from 'react'
 import warning from 'warning'
 import { useClassNames } from '../../_lib/useClassNames'
 
-import './index.css'
+import { RadioGroupContext } from './RadioGroupContext'
 
 export type RadioProps = React.PropsWithChildren<{
   value: string
@@ -64,98 +66,3 @@ const Radio = forwardRef<HTMLInputElement, RadioProps>(function RadioInner(
 })
 
 export default memo(Radio)
-
-export type RadioGroupProps<Value extends string = string> =
-  React.PropsWithChildren<{
-    className?: string
-    value?: Value
-    /**
-     * aria-label of RadioGroup
-     */
-    label?: string
-    name: string
-    onChange(next: Value): void
-    disabled?: boolean
-    readonly?: boolean
-    invalid?: boolean
-    ref?: React.Ref<HTMLDivElement>
-    'aria-labelledby'?: React.ComponentProps<'div'>['aria-labelledby']
-    'aria-orientation'?: React.ComponentProps<'div'>['aria-orientation']
-  }>
-
-interface RadioGroupContext {
-  name: string
-  selected?: string
-  disabled: boolean
-  readonly: boolean
-  invalid: boolean
-  onChange: (next: string) => void
-}
-
-const RadioGroupContext = React.createContext<RadioGroupContext>({
-  name: undefined as never,
-  selected: undefined,
-  disabled: false,
-  readonly: false,
-  invalid: false,
-  onChange() {
-    throw new Error(
-      'Cannot find onChange() handler. Perhaps you forgot to wrap with <RadioGroup> ?'
-    )
-  },
-})
-
-export const RadioGroup = forwardRef<HTMLDivElement, RadioGroupProps<string>>(
-  function RadioGroupInner(
-    {
-      value,
-      label,
-      name,
-      onChange,
-      disabled,
-      readonly,
-      invalid,
-      children,
-      ...props
-    },
-    ref
-  ) {
-    const classNames = useClassNames('charcoal-radio-group', props.className)
-
-    const handleChange = useCallback(
-      (next: string) => {
-        onChange(next)
-      },
-      [onChange]
-    )
-
-    const contextValue = useMemo(
-      () => ({
-        name,
-        selected: value,
-        disabled: disabled ?? false,
-        readonly: readonly ?? false,
-        invalid: invalid ?? false,
-        onChange: handleChange,
-      }),
-      [disabled, handleChange, invalid, name, readonly, value]
-    )
-
-    return (
-      <RadioGroupContext.Provider value={contextValue}>
-        <div
-          role="radiogroup"
-          aria-disabled={disabled}
-          aria-invalid={invalid}
-          aria-label={label}
-          aria-labelledby={props['aria-labelledby']}
-          aria-orientation={props['aria-orientation']}
-          className={classNames}
-          ref={ref}
-        >
-          {children}
-        </div>
-      </RadioGroupContext.Provider>
-    )
-  }
-) as <Value extends string>(props: RadioGroupProps<Value>) => JSX.Element
